+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## Next
+
+### ActionMenu
+
+- Action menu now uses button content.
+- ActionMenuItemContent now has bottom prop to show content below.
+  - Also has `fullWidthBottomContent` which removes the left space reserved for icons and loading states.
+
+### UseTimeoutState
+
+- Add new hook useTimeoutState.
+
+### Buttons
+
+- Add margin to right content.
+
 ## 11.2.0
 
 ### ActionMenuLink